+++ src/days/day_04.rs
@@ -1,14 +1,14 @@
 struct BingoSheet {
-    rows: Vec<Vec<u32>>,
-    columns: Vec<Vec<u32>>,
-    numbers_called_in_row: Vec<u32>,
-    numbers_called_in_column: Vec<u32>,
-    last_number_called: u32,
-    called_numbers: Vec<u32>
+    rows: Vec<Vec<usize>>,
+    columns: Vec<Vec<usize>>,
+    numbers_called_in_row: Vec<usize>,
+    numbers_called_in_column: Vec<usize>,
+    last_number_called: usize,
+    called_numbers: Vec<usize>
 }
 
 impl BingoSheet {
-    fn update(&mut self, number_called: u32) {
+    fn update(&mut self, number_called: usize) {
         for (index, row) in self.rows.iter().enumerate() {
             if row.contains(&number_called) {
                 self.numbers_called_in_row[index] += 1;
@@ -25,17 +25,17 @@ impl BingoSheet {
         self.last_number_called = number_called;
     }
 
-    fn get_result(&self) -> Option<u32> {
+    fn get_result(&self) -> Option<usize> {
         let row =  match self.numbers_called_in_row.iter().position(|count| count == &5) {
             Some(_) => {
-                let mut rsum: u32 = 0;
+                let mut rsum: usize = 0;
                 for row in &self.rows {
-                    rsum += row.iter().sum::<u32>();
+                    rsum += row.iter().sum::<usize>();
                 }
                 for num in &self.called_numbers {
                     rsum -= num;
                 }
-                Some((rsum * self.last_number_called) as u32)
+                Some((rsum * self.last_number_called) as usize)
             },
             None => None
         };
@@ -45,14 +45,14 @@ impl BingoSheet {
         } else {
             return match self.numbers_called_in_column.iter().position(|count| count == &5) {
                 Some(_) => {
-                    let mut rsum: u32 = 0;
+                    let mut rsum: usize = 0;
                     for row in &self.rows {
-                        rsum += row.iter().sum::<u32>();
+                        rsum += row.iter().sum::<usize>();
                     }
                     for num in &self.called_numbers {
                         rsum -= num;
                     }
-                    Some((rsum * self.last_number_called) as u32)
+                    Some((rsum * self.last_number_called) as usize)
                 },
                 None => None
             };
@@ -60,20 +60,20 @@ impl BingoSheet {
     }
 }
 
-pub(crate) fn a(input: &str) -> u32 {
+pub(crate) fn a(input: &str) -> usize {
     let mut bingo_lines: Vec<&str> = input.split("\n\n").collect();
-    let chosen_numbers: Vec<u32> = bingo_lines[0].split(",").map(|num| num.parse().unwrap()).collect();
+    let chosen_numbers: Vec<usize> = bingo_lines[0].split(",").map(|num| num.parse().unwrap()).collect();
 
     let mut sheets: Vec<BingoSheet> = Vec::new();
     bingo_lines.remove(0);
 
     for sheet in bingo_lines {
-        let rows: Vec<Vec<u32>> = sheet.split("\n")
+        let rows: Vec<Vec<usize>> = sheet.split("\n")
             .map(|row| row.split(" ").filter(|splitres| splitres != &"")
                 .map(|num| num.parse().unwrap()).collect())
             .collect();
 
-        let mut columns: Vec<Vec<u32>> = vec![Vec::new(); rows.len()];
+        let mut columns: Vec<Vec<usize>> = vec![Vec::new(); rows.len()];
         for row in &rows {
             for (index, num) in row.iter().enumerate() {
                 columns[index].push(*num);
@@ -103,20 +103,20 @@ pub(crate) fn a(input: &str) -> u32 {
     unreachable!();
 }
 
-pub(crate) fn b(input: &str) -> u32 {
+pub(crate) fn b(input: &str) -> usize {
     let mut bingo_lines: Vec<&str> = input.split("\n\n").collect();
-    let chosen_numbers: Vec<u32> = bingo_lines[0].split(",").map(|num| num.parse().unwrap()).collect();
+    let chosen_numbers: Vec<usize> = bingo_lines[0].split(",").map(|num| num.parse().unwrap()).collect();
 
     let mut sheets: Vec<BingoSheet> = Vec::new();
     bingo_lines.remove(0);
 
     for sheet in bingo_lines {
-        let rows: Vec<Vec<u32>> = sheet.split("\n")
+        let rows: Vec<Vec<usize>> = sheet.split("\n")
             .map(|row| row.split(" ").filter(|splitres| splitres != &"")
                 .map(|num| num.parse().unwrap()).collect())
             .collect();
 
-        let mut columns: Vec<Vec<u32>> = vec![Vec::new(); rows.len()];
+        let mut columns: Vec<Vec<usize>> = vec![Vec::new(); rows.len()];
         for row in &rows {
             for (index, num) in row.iter().enumerate() {
                 columns[index].push(*num);
